refactor: 대기 소켓, 스키마 이름 변경 (auto-join -> auto-game)
