+++ src/v1/sockets/waiting/schemas/auto-game.schema.ts
@@ -1,8 +1,8 @@
 import { TypeOf, z } from 'zod';
 import { tournamentSizeSchema } from './tournament.schema.js';
 
-export type customJoinSchemaType = TypeOf<typeof autoJoinSchema>;
+export type customJoinSchemaType = TypeOf<typeof autoGameSchema>;
 
-export const autoJoinSchema = z.object({
+export const autoGameSchema = z.object({
   tournamentSize: tournamentSizeSchema,
 });
